+++ packages/react/src/components/Breadcrumbs/Breadcrumbs.tsx
@@ -6,19 +6,26 @@ import type { BreadcrumbsProps } from './Breadcrumbs.props';
 import { withGlobalPrefix } from '../../helpers/with-global-prefix';
 import { extractProps } from '../../helpers/extract-props';
 import { marginPropDefs } from '../../props/margin.props';
+import { BreadcrumbsContext } from './Breadcrumbs.context';
 
 const COMPONENT_NAME = 'Breadcrumbs';
 const componentClassName = withGlobalPrefix(COMPONENT_NAME);
 
 export const Breadcrumbs: React.FC<BreadcrumbsProps> = props => {
-  const { className, children, ...breadcrumbsProps } = extractProps(props, marginPropDefs);
+  const { className, children, inverted, ...breadcrumbsProps } = extractProps(
+    props,
+    marginPropDefs
+  );
   return (
     <nav
       aria-label="breadcrumbs"
       className={clsx(componentClassName, className)}
+      data-inverted={inverted ? '' : undefined}
       {...breadcrumbsProps}
     >
-      <ol role="list">{children}</ol>
+      <ol role="list">
+        <BreadcrumbsContext.Provider value={{ inverted }}>{children}</BreadcrumbsContext.Provider>
+      </ol>
     </nav>
   );
 };