+++ packages/react/src/types/css-token-variable.ts
@@ -1,0 +1,1 @@
+export type CssTokenVariable = `var(--h-${string})`;